Dev cdrobaina01

+++ sql/scripts/init_02_1_functions_delete.sql
@@ -1,31 +1,38 @@
-﻿-- DELETE FUNCTIONS
-CREATE OR REPLACE FUNCTION delete_fee(text) RETURNS void AS $$
+﻿-- ADMIN DELETE FUNCTIONS
+CREATE OR REPLACE FUNCTION delete_fee_admin(text) RETURNS void AS $$
 BEGIN
 	DELETE FROM fee WHERE name = $1;
 END; $$
 LANGUAGE plpgsql;
 
-CREATE OR REPLACE FUNCTION delete_rol(integer) RETURNS void AS $$
+CREATE OR REPLACE FUNCTION delete_rol_admin(integer) RETURNS void AS $$
 BEGIN
 	DELETE FROM rol WHERE id = $1;
 END; $$
 LANGUAGE plpgsql;
 
-CREATE OR REPLACE FUNCTION delete_user(integer) RETURNS void AS $$
-BEGIN
-	DELETE FROM user_local WHERE id = $1;
-END; $$
-LANGUAGE plpgsql;
-
+-- SPECIAL
 CREATE OR REPLACE FUNCTION delete_situation(text) RETURNS void AS $$
 BEGIN
-	DELETE FROM situation WHERE name = $1;
+	IF $1 NOT IN ('disponible', 'alquilado', 'en taller') THEN
+		DELETE FROM situation WHERE name = $1;
+	END IF;
 END; $$
 LANGUAGE plpgsql;
 
 CREATE OR REPLACE FUNCTION delete_paymethod(text) RETURNS void AS $$
 BEGIN
-	DELETE FROM pay_method WHERE name = $1;
+	IF $1 NOT IN ('efectivo', 'cheque', 'tarjeta de crédito') THEN
+		DELETE FROM pay_method WHERE name = $1;
+	END IF;
+END; $$
+LANGUAGE plpgsql;
+
+
+-- REGULAR DELETE FUNCTIONS
+CREATE OR REPLACE FUNCTION delete_user(integer) RETURNS void AS $$
+BEGIN
+	DELETE FROM user_local WHERE id = $1;
 END; $$
 LANGUAGE plpgsql;
